修改班级管理 更新数据库

+++ src/main/java/com/tjh/service/impl/TeacherServiceImpl.java
@@ -2,11 +2,14 @@ package com.tjh.service.impl;
 
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
+import com.tjh.constant.Constant;
 import com.tjh.mapper.TeacherMapper;
 import com.tjh.pojo.Teacher;
+import com.tjh.service.SysRoleService;
 import com.tjh.service.TeacherService;
 import com.tjh.util.DataGridView;
 import com.tjh.util.PageBuilder;
+import com.tjh.util.ResultMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,6 +22,9 @@ import java.util.Map;
 public class TeacherServiceImpl implements TeacherService {
     @Autowired
     private TeacherMapper teacherMapper;
+    @Autowired
+    private SysRoleService roleService;
+
     @Override
     public Teacher loadOneTeacherByClassId(Integer classId) {
         return  teacherMapper.loadOneTeacherByClassId(classId);
@@ -44,19 +50,21 @@ public class TeacherServiceImpl implements TeacherService {
 
     @Override
     public int addTeacher(Teacher teacher) {
-        int i = this.teacherMapper.addTeacher(teacher);
-//        int temp = 0;
-//        if(i>0){
-//            Map<String,Integer> map = new HashMap<>();
-//            map.put("userId",userId);
-//            map.put("teacherId",teacher.getTeacherId());
-//            temp = this.teacherMapper.addTeacherUserInfo(map);
-//        }
+        ResultMessage resultMessage = this.roleService.addRoleUser(Constant.TEACHER_ROLE_ID, teacher.getTeacherId());
+        int i = 0;
+        if(resultMessage.getCode() == 200){
+            i = this.teacherMapper.addTeacher(teacher);
+        }
         return  i;
     }
 
     @Override
     public int deleteTeacherStateByTeacherId(Integer teacherId) {
-        return this.teacherMapper.deleteTeacherStateByTeacherId(teacherId);
+        ResultMessage resultMessage = roleService.deleteRoleUser(Constant.TEACHER_ROLE_ID, teacherId);
+        int i = 0;
+        if(resultMessage.getCode() == 200){
+            i = this.teacherMapper.deleteTeacherStateByTeacherId(teacherId);
+        }
+        return i;
     }
 }
